Packages: add support for Windows
