+++ src/fastfetch.c
@@ -124,7 +124,7 @@ static inline void printCommandHelp(const char* command)
     }
     else if(strcasecmp(command, "packages-format") == 0)
     {
-        constructAndPrintCommandHelpFormat("packages", "{2} (pacman){?3}[{3}]{?}, {4} (dpkg), {5} (rpm), {6} (emerge), {7} (xbps), {8} (nix-system), {9} (nix-user), {10} (nix-default), {11} (apk), {12} (pkg), {13} (flatpak), {14} (snap), {15} (brew), {16} (port)", 16,
+        constructAndPrintCommandHelpFormat("packages", "{2} (pacman){?3}[{3}]{?}, {4} (dpkg), {5} (rpm), {6} (emerge), {7} (xbps), {8} (nix-system), {9} (nix-user), {10} (nix-default), {11} (apk), {12} (pkg), {13} (flatpak), {14} (snap), {15} (brew), {16} (port), {17} (scoop)", 17,
             "Number of all packages",
             "Number of pacman packages",
             "Pacman branch on manjaro",
@@ -141,6 +141,7 @@ static inline void printCommandHelp(const char* command)
             "Number of snap packages",
             "Number of brew packages",
             "Number of macports packages"
+            "Number of scoop packages"
         );
     }
     else if(strcasecmp(command, "shell-format") == 0)
